% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binarify.R
\name{binarify}
\alias{binarify}
\title{Binarify}
\usage{
binarify(data, verbose = FALSE, copy_data = FALSE)
}
\arguments{
\item{data}{data.frame containing the original voter file.}

\item{verbose}{boolean whether to print extra stuff.}

\item{copy_data}{boolean whether to force a copy of the original data}
}
\value{
data.table -- either by reference (default) or value (if \code{copy_data = TRUE})
}
\description{
Function to convert character columns with 2 values to 0L and 1L.
}
\details{
Function to convert all character columns of a data frame with two unique values
where one is equal to 'y', 'yes', or 'support' to 1L and all blanks to 0L
}
\examples{
#df = readRDS("/media/aaron/DATA/VF/ut.rds")
#binarify(df, verbose=T) # Note that by default, it edits the data by reference, no copy needed
#df = binarify(df, T)
#dt = binarify(df, T, copy_data = T) # this will force it to make a copy of the original data
}

