% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bins.R
\name{bins}
\alias{binify}
\alias{bins}
\title{Assign predicted probabilities to bins}
\usage{
bins(probs, obs, nm = "predicted")

binify(x, breaks = 20)
}
\arguments{
\item{probs}{Predicted probabilities (0 to 1)}

\item{obs}{observed outcomes (0 or 1)}

\item{nm}{Output column name}
}
\value{
dataframe with probs and bins
}
\description{
Assign predicted probabilities to bins
}
\section{Functions}{
\itemize{
\item \code{binify}: 
}}

