% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check-read.R
\name{check_read}
\alias{check_read}
\alias{fileSize}
\alias{objectSize}
\title{Check segment reads data prior to sending to vendor}
\usage{
check_read(obs_path, targ_path, obs_path_last = NULL)

objectSize(x)

fileSize(x)
}
\arguments{
\item{obs_path}{Directory path to folder where call lists are located}

\item{targ_path}{Path to calc.csv file}

\item{obs_path_last}{Optional directory path to folder where previous weeks' call lists are located}
}
\value{
data frame of aggregated call list
}
\description{
Check segment reads data prior to sending to vendor
}
\section{Functions}{
\itemize{
\item \code{objectSize}: Object size nicely formatted

\item \code{fileSize}: File size in "MB"
}}

