% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean-names.R
\name{clean_names}
\alias{clean_names}
\alias{shorten_names}
\title{Clean names}
\usage{
clean_names(x)

shorten_names(x)
}
\arguments{
\item{string}{names to be cleaned}
}
\value{
string cleaned names
}
\description{
Ensures that column names are valid R names
}
\details{
Replaces hyphens with underscores and adds a 'v' before columns
that start with numbers. Removes all non-alphanumeric or underscore
}
\section{Functions}{
\itemize{
\item \code{shorten_names}: 
}}
\examples{
# setnames(df, names(df), clean_names(names(df)))
}

