% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_dir}
\alias{create_dir}
\title{Create a new subdirectory}
\usage{
create_dir(mainDir, subDir = NULL)
}
\arguments{
\item{mainDir}{Path to the directory to which you'd like to add a subdirectory}

\item{subDir}{Name of the subdirectory}
}
\description{
Create a new subdirectory
}

