% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delete-all-missing.R
\name{delete_all_missing}
\alias{any_missing}
\alias{delete_all_missing}
\title{Delete all missing}
\usage{
delete_all_missing(data, delete = TRUE)

any_missing(x)
}
\arguments{
\item{data}{data.table containing columns to be deleted}

\item{delete}{boolean -- whether or not to actually delete the columns}
}
\value{
data.table reference (it does not pass by value remember)
}
\description{
Delete any columns with all missing values
}
\details{
This function passes the data table by reference, so be warned that
\code{data} will always be changed. If you run with delete = FALSE,
it will list out the columns to be deleted, but will not delete them.
This is advised while testing out regular expressions.
}
\section{Functions}{
\itemize{
\item \code{any_missing}: 
}}
\examples{
#vf = delete_all_missing(vf, "electionreturns")
#vf = delete_all_missing(vf, "Idon'tknow", delete = FALSE)
}

