% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delete-by-regex.R
\name{delete_by_regex}
\alias{delete_by_regex}
\title{Delete by regex}
\usage{
delete_by_regex(data, regex, delete = TRUE)
}
\arguments{
\item{data}{data.table containing columns to be deleted}

\item{regex}{string containing regular expression to match for column names}

\item{delete}{boolean -- whether or not to actually delete the columns}
}
\value{
data.table reference (it does not pass by value remember)
}
\description{
Delete all columns in a data table that match a particular regular expression
}
\details{
This function passes the data table by reference, so be warned that
\code{data} will always be changed. If you run with delete = FALSE, it will
list out the columns to be deleted, but will not delete them. This is
advised while testing out regular expressions.
}
\examples{
#vf = delete_by_regex(vf, "electionreturns")
#vf = delete_by_regex(vf, "Idon'tknow", delete=FALSE)
}

