% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delete-zero-var.R
\name{delete_zero_var}
\alias{delete_zero_var}
\title{Delete zero var}
\usage{
delete_zero_var(data, delete = TRUE)
}
\arguments{
\item{data}{data.table containing columns to be deleted}

\item{delete}{boolean -- whether or not to actually delete the columns}
}
\value{
data.table reference (it does not pass by value remember)
}
\description{
Delete all columns in a data table that have one unique value
}
\details{
This function passes the data table by reference, so be warned that
\code{data} will always be changed. If you run with delete = FALSE, it will
list out the columns to be deleted, but will not delete them. This is
advised while testing out regular expressions.
}
\examples{
#vf = delete_zero_var(vf)
#vf = delete_zero_var(vf, delete=FALSE)
}

