% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display-cm.R
\name{display_cm}
\alias{.format_cm}
\alias{display_buckets}
\alias{display_cm}
\alias{table_out}
\title{Display confusion matrix}
\usage{
display_cm(results, positive = "yes")

.format_cm(label, cM, obs, probs)

display_buckets(results, positive = "yes", negative = "no")

table_out(cmat, buckets, title = "Model Performance Comparison")
}
\arguments{
\item{results}{a results file created using the get_results() function}

\item{positive}{value of outcome that should be classified as positive;
default is 'yes'}

\item{label}{A label (e.g., 'lda') for the model}

\item{cM}{A confusion matrix}

\item{negative}{Label for negative class; default = 'no'}

\item{cmat}{Confusion matrix}

\item{buckets}{}

\item{results}{A results file created using get_results() function}

\item{positive}{Label for positive class; default = 'yes'}
}
\value{
a formatted list of models with accompanying test statistics

A dataframe with one row

a data frame with percentages and counts for each model for each bin
}
\description{
Display confusion matrix

Internal function to format confusion matrix results

Function to bin results into 5% buckets

Title of html file
}
\section{Functions}{
\itemize{
\item \code{.format_cm}: 

\item \code{display_buckets}: 

\item \code{table_out}: 
}}

