% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatters}
\alias{formatters}
\alias{pN}
\alias{totitle}
\title{Pretty Num with a comma as big.mark}
\usage{
formatters(x)

pN(x)

totitle(x)
}
\description{
Pretty Num with a comma as big.mark
}
\section{Functions}{
\itemize{
\item \code{}: 

\item \code{}: 
}}

