% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/group-ages.R
\name{group_ages}
\alias{group_ages}
\title{Group into our standard age groups}
\usage{
group_ages(x)
}
\arguments{
\item{x}{vector of ages}
}
\value{
grouped ages
}
\description{
Group into our standard age groups
}

