% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/group-ethnicities.R
\name{group_ethnicities}
\alias{group_ethnicities}
\title{Group into standard ethnicities}
\usage{
group_ethnicities(x)
}
\arguments{
\item{x}{vector of ethnicities}
}
\value{
grouped ethnicities
}
\description{
Group into standard ethnicities
}

