% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Installs a list of packages}
\usage{
install_packages(packages = NULL)
}
\arguments{
\item{packages}{A list of packages, or by default my list.}
}
\value{
\code{NULL}
}
\description{
Installs a list of packages
}

