% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lclip.R
\name{lclip}
\alias{lclip}
\title{Export data frame to linux clipboard}
\usage{
lclip(x, sep = "\\t", row.names = FALSE, col.names = TRUE)
}
\arguments{
\item{x}{data frame to export}

\item{sep}{Seperator to use (default = \code{"\t"})}

\item{row.names}{Boolean -- whether to save row names (default = \code{FALSE})}

\item{col.names}{Boolean -- whether to save column names (default = \code{TRUE})}
}
\description{
Export data frame to linux clipboard
}

