% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mc-near-zero-variance.R
\name{mcNearZeroVar}
\alias{mcImpute}
\alias{mcNearZeroVar}
\title{Use multiple cores to test for variables with near-zero-variance.}
\usage{
mcNearZeroVar(x, freqCut = 95/5, uniqueCut = 10, saveMetrics = FALSE)

mcImpute(x)
}
\arguments{
\item{x}{Data frame}

\item{freqCut}{Frequency threshold for being considered low variance.}

\item{uniqueCut}{Threshold for minimum number of unique values}

\item{saveMetrics}{Whether to return metrics or just a list of variables}
}
\value{
Variables with zero or near-zero variance or corresponding metrics.
}
\description{
Use multiple cores to test for variables with near-zero-variance.
}
\section{Functions}{
\itemize{
\item \code{mcImpute}: 
}}

