% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pre_process}
\alias{.factorize}
\alias{pre_process}
\title{Preprocess data for modeling}
\usage{
pre_process(universe, impute = TRUE, freqRatioLevel = 100,
  corCutOff = 0.99, dummy = FALSE, zscore = FALSE)

.factorize(x, freqCut = NULL, percentCut = NULL, otherFactor = TRUE)
}
\arguments{
\item{universe}{Dataframe}

\item{impute}{Default = TRUE}

\item{freqRatioLevel}{Default = 100}

\item{corCutOff}{Default = 0.99}

\item{dummy}{Default = FALSE}

\item{zscore}{Default = FALSE}

\item{x}{Data}

\item{freqCut}{Minimum cases per level of the factor to still include}

\item{percentCut}{Minimum percentage cases within a level to still include}

\item{otherFactor}{Relabel excluded factor levels as "other"}
}
\description{
Preprocess data for modeling

A Function to help with creating factors in R
}
\section{Functions}{
\itemize{
\item \code{.factorize}: 
}}

