% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict-outcomes.R
\name{predict_outcomes}
\alias{predict_outcomes}
\title{Predict outcomes}
\usage{
predict_outcomes(tune_list, train_data, test_data, outcome = "Class",
  outcome_level = "yes")
}
\arguments{
\item{tune_list}{List of model training outputs e.g.,
\code{tune_list = list(lda = lda_tune, lr = lr_tune, bn = bn_tune)}}

\item{train_data}{Training data set as data.frame}

\item{test_data}{Test data set as data.frame.}

\item{outcome}{Name of the dependent variable 'outcome' by default}

\item{outcome_level}{level of interest in the DV}
}
\value{
List consisting of two data.frames labeled "train_results" and
"test_results"
}
\description{
Aggregated results from a list of tuning outputs from various models
}

