% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare-universe.R
\name{prepare_universe}
\alias{prepare_universe}
\alias{prepare_universes}
\title{Prepare universe for modeling}
\usage{
prepare_universe(universe, class_name, save_dir, list_of_models = NULL)

prepare_universes(config)
}
\arguments{
\item{universe}{data frame containing variable to be predicted
\code{class_name} along with an id variable and all modeling variables.}

\item{class_name}{character string of variable to be predicted -- should
have two levels \code{c(0, 1)}.}

\item{save_dir}{character path of directory into which the training universe
and testing universes will be saved.}

\item{list_of_models}{list of names of models to be removed when
\code{prepare_universe} is being run on several models.
These variables are removed from \code{universe} prior to modeling.}

\item{config}{}

\item{id}{character vector of the id column. This defaults to
\code{"id"}, but could be changed if necessary.}
}
\description{
This function takes care of preparing a universe for a particular model with
the model's name equal to \code{class_name}, which incidentally is the
class name -- go figure. Key actions that take place within this function are
to filter unwanted variables, split the universe into training and test sets,
and save them in a standardized directory (i.e., \code{save_dir}).
}
\section{Functions}{
\itemize{
\item \code{prepare_universes}: 
}}
\examples{
# prepare_universe(df, "unsure", "/Dropbox/output_dir/")
}
\seealso{
\link{train_universe}
}
\keyword{modeling}
\keyword{universe}

