% Generated by roxygen2 (4.0.2): do not edit by hand
\name{println}
\alias{println}
\title{Like cat except with a newline character at the end and default sep of ""}
\usage{
println(...)
}
\arguments{
\item{...}{Character expressions to be printed to screen}
}
\description{
Like cat except with a newline character at the end and default sep of ""
}

