% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read-dir.R
\name{read_dir}
\alias{read_dir}
\title{Read all (csv/tab/txt) files in a directory}
\usage{
read_dir(x, combine = TRUE)
}
\arguments{
\item{combine}{Boolean, whether to return one data.frame or a list of
data.frames}

\item{path}{Directory path}
}
\value{
Either a single aggregated data frame (if \code{combined = T}
[default]) or a list of data frames (if \code{combined = F}).
}
\description{
Read all (csv/tab/txt) files in a directory
}

