% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read-google.R
\name{readGoogle}
\alias{readGoogle}
\title{Read google spreadsheet}
\usage{
readGoogle(url, sheet = 1)
}
\arguments{
\item{url}{of exported spreadsheet}

\item{sheet}{number}
}
\value{
google spreadsheet in the form of a data frame
}
\description{
Read google spreadsheet
}

