% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/result-predictions.R
\name{result_predictions}
\alias{compute_metrics}
\alias{metricify}
\alias{result_predictions}
\title{Result predictions}
\usage{
result_predictions(m, config)

metricify(m, config)

compute_metrics(config)
}
\arguments{
\item{m}{}

\item{config}{}

\item{config}{}
}
\description{
Result predictions
}
\section{Functions}{
\itemize{
\item \code{metricify}: 

\item \code{compute_metrics}: 
}}

