% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run-models.R
\name{run_models}
\alias{append_paths_to_config}
\alias{create.dir}
\alias{fix_slashes}
\alias{merge_by_key}
\alias{run_models}
\alias{safe_paths}
\title{Run models}
\usage{
run_models(config_path)

create.dir(mainDir, subDir = NULL)

fix_slashes(path)

safe_paths(...)

merge_by_key(large, small, key_name = "id", mult = "first")

append_paths_to_config(config)
}
\arguments{
\item{config_path}{Location of yaml config file}

\item{mainDir}{}

\item{subDir}{(Default = NULL)}
}
\description{
Run models
}
\section{Functions}{
\itemize{
\item \code{create.dir}: 

\item \code{fix_slashes}: 

\item \code{safe_paths}: 

\item \code{merge_by_key}: 

\item \code{append_paths_to_config}: 
}}

