% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score-models.R
\name{score_models}
\alias{score_models}
\alias{score_models_split}
\title{Score models}
\usage{
score_models(config)

score_models_split(config, models)
}
\arguments{
\item{config}{}

\item{models}{}

\item{model_name}{}

\item{config}{}
}
\description{
Score models
}
\section{Functions}{
\itemize{
\item \code{score_models_split}: 
}}

