% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrub-income.R
\name{scrub_income}
\alias{scrub_income}
\title{Scrub income}
\usage{
scrub_income(x)
}
\arguments{
\item{string}{of original variable}
}
\value{
double scaled and centered (internally)
}
\description{
Cleans up CommercialData_EstimatedIncome variable
}

