% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/source-dir.R
\name{source_dir}
\alias{source_dir}
\title{Source all files in a director}
\usage{
source_dir(path = "R/")
}
\arguments{
\item{path}{Directory path (default = \code{"R/"})}
}
\description{
Source all files in a director
}

