% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timed-try.R
\name{timed_try}
\alias{timed_try}
\title{Timed try function}
\usage{
timed_try(expr, cpu = Inf, elapsed = Inf)
}
\arguments{
\item{expr}{Expression to be wrapped in try function}

\item{cpu}{(default=Inf)}

\item{elapsed}{(default=Inf)}
}
\value{
NULL or results of \code{expr}
}
\description{
Timed try function
}

