% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/train-universe.R
\name{train_universe}
\alias{train_model}
\alias{train_universe}
\alias{train_universes}
\title{Train a universe using a variety of models}
\usage{
train_universe(m, config)

train_universes(config)

train_model(condition, method, libs = NULL, tuneGrid = NULL)
}
\arguments{
\item{config}{list containing configuration instructs
(e.g., which models to run, where to save models, etc.).}

\item{model}{character name of model previously
prepped with \code{prepare_universe()}}
}
\description{
This function is where the actual modeling takes place.
Models are specified in a \code{config} file that acts as
a sort of controller for the modeling process.
}
\section{Functions}{
\itemize{
\item \code{train_universes}: 

\item \code{train_model}: 
}}
\examples{
#train_universe("unsure", config)
#train_universe("turnout", config)
}
\seealso{
\link{prepare_universe}
}
\keyword{model}
\keyword{modeling}
\keyword{universe}

