% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vswitch}
\alias{vswitch}
\title{Switch for vectors like ifelse}
\usage{
vswitch(x, ...)
}
\arguments{
\item{x}{Vector of values on which to apply the switch}

\item{...}{switch expressions}
}
\value{
vector of new values
}
\description{
Switch for vectors like ifelse
}

