% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/winner.R
\name{winner}
\alias{winner}
\alias{winners}
\title{Pick winner}
\usage{
winner(model_name, config)

winners(config)
}
\arguments{
\item{model_name}{}

\item{config}{}

\item{config}{}
}
\description{
Pick winner
}
\section{Functions}{
\itemize{
\item \code{winners}: 
}}

