% Generated by roxygen2 (4.0.2): do not edit by hand
\name{with_leaners}
\alias{with_leaners}
\title{Top line with leaners}
\usage{
with_leaners(x, w, q1 = "Q2", q2 = "Q3")
}
\arguments{
\item{x}{segment data}

\item{w}{weights}

\item{q1}{character vector of length 1 (e.g., q1 = "Q2")}

\item{q2}{character vector of length 1 (e.g., q2 = "Q2")}
}
\value{
overall read with leaners
}
\description{
Top line with leaners
}

