% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Balanced segment read}
\usage{
balance(q, x, w)
}
\arguments{
\item{q}{Character vector (e.g., "Q1")}

\item{x}{segment data}

\item{w}{weights}
}
\value{
balenced read
}
\description{
Balanced segment read
}

