% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_read}
\alias{check_read}
\alias{ob.size}
\title{Check segment reads prior to sending them}
\usage{
check_read(obs_path, targ_path, obs_path_last = NULL)

ob.size(x)
}
\arguments{
\item{obs_path}{Directory path to folder where call lists are located}

\item{targ_path}{Path to calc.csv file}

\item{obs_path_last}{Optional directory path to folder where previous weeks' call lists are located}
}
\value{
data frame of aggregated call list
}
\description{
Check segment reads prior to sending them
}

