% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatters}
\alias{formatters}
\alias{pN}
\alias{totitle}
\title{Pretty Num with a comma as big.mark}
\usage{
formatters(x)

pN(x)

totitle(x)
}
\description{
Pretty Num with a comma as big.mark
}
\section{Functions}{
\itemize{
\item \code{pN}: 

\item \code{totitle}: 
}}

