% Generated by roxygen2 (4.0.2): do not edit by hand
\name{group_ethnicities}
\alias{group_ethnicities}
\title{Group into standard ethnicities}
\usage{
group_ethnicities(x)
}
\arguments{
\item{x}{vector of ethnicities}
}
\value{
grouped ethnicities
}
\description{
Group into standard ethnicities
}

