% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\title{Regex to see if vector contains value}
\usage{
has(x, regex)
}
\arguments{
\item{x}{vector to be evaluated for value}

\item{regex}{expression}
}
\value{
Boolean, whether or not the vector contains the value
}
\description{
Regex to see if vector contains value
}

