% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/println.R
\name{println}
\alias{println}
\title{Like cat except with a newline character at the end and default sep of ""}
\usage{
println(...)
}
\arguments{
\item{...}{Character expressions to be printed to screen}
}
\description{
Like cat except with a newline character at the end and default sep of ""
}

