% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read-any.R
\name{read_any}
\alias{pN}
\alias{read.all}
\alias{read_any}
\title{Read any common file depending upon file extension}
\usage{
read_any(x, to_keep = NULL)

read.all(x, to_keep = NULL)

pN(x)
}
\arguments{
\item{x}{Path to file}

\item{to_keep}{Character vector of variables to keep (default = \code{NULL})}
}
\value{
data frame
}
\description{
Read any common file depending upon file extension
}
\section{Functions}{
\itemize{
\item \code{read.all}: 

\item \code{pN}: 
}}

