% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_dir}
\alias{read_dir}
\title{Read all (csv/tab/txt) files in a directory}
\usage{
read_dir(path, combine = T)
}
\arguments{
\item{path}{Directory path}

\item{combine}{Boolean, whether to return one data.frame or a list of data.frames}
}
\value{
Either a single aggregated data frame (if \code{combined = T} [default]) or a list of data frames (if \code{combined = F}).
}
\description{
Read all (csv/tab/txt) files in a directory
}

